Time: 69 ms (6.67%) | Memory: 48.2 MB (17.33%) - LeetSync

+++ 3974-xor-after-range-multiplication-queries-i/xor-after-range-multiplication-queries-i.java
@@ -1,6 +1,6 @@
 class Solution {
     public int xorAfterQueries(int[] nums, int[][] queries) {
-        long a =1_000_000_007;
+        int a =1_000_000_007;
         for(int i=0;i<queries.length;i++)
         {
             for(int j=queries[i][0];j<=queries[i][1];j+=queries[i][2])
